% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_k.R
\name{A_k}
\alias{A_k}
\title{A_k: The function that returns the matrix A_k}
\usage{
A_k(k)
}
\arguments{
\item{k}{Natural number, the precision with which to approximate squares
within \eqn{[0,1]}}
}
\value{
A neural network that approximates the square of any real within
\eqn{[0,1]}
}
\description{
A_k: The function that returns the matrix A_k
}
