% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aff.R
\name{Aff}
\alias{Aff}
\title{Aff}
\usage{
Aff(W, b)
}
\arguments{
\item{W}{An \eqn{m \times n} matrix representing the weight of the affine
neural network}

\item{b}{An \eqn{m \times 1} vector representing the bias of the affine
neural network}
}
\value{
Returns the network \eqn{((W,b))} representing an affine neural network. Also
denoted as \eqn{\mathsf{Aff}_{W,b}}
See also \code{\link{Cpy}} and \code{\link{Sum}}.
}
\description{
The function that returns \eqn{\mathsf{Aff}} neural networks.
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al. Space-time error estimates for deep
neural network approximations for differential equations. Adv Comput Math 49, 4 (2023).
https://doi.org/10.1007/s10444-022-09970-2

Definition 2.3.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}
}
