% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_k.R
\docType{data}
\name{B}
\alias{B}
\title{This is an intermediate variable, see reference.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 1 columns.
}
\usage{
B
}
\description{
This is an intermediate variable, see reference.
}
\keyword{datasets}
