% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_k.R
\name{C_k}
\alias{C_k}
\title{C_k: The function that returns the C_k matrix}
\usage{
C_k(k)
}
\arguments{
\item{k}{Natural number, the precision with which to approximate squares
within \eqn{[0,1]}}
}
\value{
A neural network that approximates the square of any real within
\eqn{[0,1]}
}
\description{
C_k: The function that returns the C_k matrix
}
