% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\encoding{utf8}
\name{comp}
\alias{comp}
\alias{\%comp\%}
\title{comp}
\usage{
comp(phi_1, phi_2)

phi_1 \%comp\% phi_2
}
\arguments{
\item{phi_1}{first neural network to be composed, goes on the left}

\item{phi_2}{second neural network to be composed, goes on right}
}
\value{
The composed neural network. See also \code{\link{dep}}.
Composition of neural networks is the operation defined for \eqn{\nu_1 \in \mathsf{NN}}
and \eqn{\nu_2 \in \mathsf{NN}} as:

\deqn{
\nu_1 \bullet \nu_2 = \begin{cases} (( W'_1,b'_1 ),
( W'_2,b'_2 ), ...,( W'_{M-1}, b'_{M-1}), ( W_1W'_M, W_1b'_{M} + b_1), (W_2, b_2 ),\\...,
( W_L,b_L )) & :( L> 1 ) \land ( M > 1 ) \\((W_1W'_1,W_1b'_1+b_1),(W_2,b_2), (W_3,b_3),...,
(W_Lb_L)) & :(L>1) \land (M=1) \\((W'_1, b'_1),(W'_2,b'_2), ...,
(W'_{M-1}, b'_{M-1})(W_1, b'_M + b_1)) &:(L=1) \land (M>1) \\ ((W_1W'_1, W_1b'_1+b_1)) &:(L=1)
\land (M=1)\end{cases}

}
}
\description{
The function that takes the composition of two neural
networks assuming they are compatible, i.e., given
\eqn{\nu_1, \nu_2 \in \mathsf{NN}}, it must be the case that
\eqn{\mathsf{I}(\nu)_1 = \mathsf{O}(\nu_2)}.
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al.
Space-time error estimates for deep neural network approximations
for differential equations. Adv Comput Math 49, 4 (2023).
\url{https://rdcu.be/dxYVV}

Specifically:

Definition 2.1.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}

\emph{Remark:} We have two versions of this function, an
infix version for close resemblance to mathematical notation and
prefix version.
}
