% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aff.R
\name{Cpy}
\alias{Cpy}
\title{Cpy}
\usage{
Cpy(n, k)
}
\arguments{
\item{n}{number of copies to make.}

\item{k}{the size of the input vector.}
}
\value{
Returns an affine network that makes a concatenated vector that is \eqn{n}
copies of the input vector of size \eqn{k}. See \code{\link{Aff}} and \code{\link{Sum}}.
}
\description{
The function that returns \eqn{\mathsf{Cpy}} neural networks.
These are neural networks defined as such
\deqn{
\mathsf{Aff}_{\left[ \mathbb{I}_k \: \mathbb{I}_k \: \cdots \: \mathbb{I}_k\right]^T,0_{k}}
}
}
\references{
Definition 2.9. Rafi S., Padgett, J.L., Nakarmi, U. (2024) Towards an
Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
