% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{create_block_diagonal}
\alias{create_block_diagonal}
\title{Function for creating a block diagonal given two matrices.}
\usage{
create_block_diagonal(matrix1, matrix2)
}
\arguments{
\item{matrix1}{A matrix.}

\item{matrix2}{A matrix}
}
\value{
A block diagonal matrix with matrix1 on top left
and matrix2 on bottom right.
}
\description{
Function for creating a block diagonal given two matrices.
}
