% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{dep}
\alias{dep}
\title{dep}
\usage{
dep(nu)
}
\arguments{
\item{nu}{a neural network of the type generated by
create_nn(). Very straightforwardly it is the
length of the list where neural networks are defined as an odered list of
lists.}
}
\value{
Integer representing the depth of the neural network.
}
\description{
The function that returns the depth of a neural network. Denoted
\eqn{\mathsf{D}}.
}
\examples{
create_nn(c(4, 5, 6, 2)) |> dep()
}
\references{
Definition 1.3.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}.
}
