% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etr.R
\name{Etr}
\alias{Etr}
\title{Etr}
\usage{
Etr(n, h)
}
\arguments{
\item{n}{number of trapezoids to make. Note this will result in a set of
trapezoids.
Note that this will result in n+1 meshpoints including the starting a and
ending b

\emph{Note: } Upon instantiation with any continuous function this neural
network must be fed with \eqn{n+1} real numbers representing the values
of the function being approximated at the \eqn{n+1} meshpoints which are
the legs of the \eqn{n} triangles as stipulated in the input parameters.}

\item{h}{width of trapezoids}
}
\value{
an approximation for value of the integral.
}
\description{
The function that returns the \eqn{\mathsf{Etr}} networks.
}
\references{
Definition 2.33. Rafi S., Padgett, J.L., Nakarmi, U. (2024)
Towards an Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
