% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_creator.R
\name{generate_random_matrix}
\alias{generate_random_matrix}
\title{Function to generate a random matrix with specified dimensions.}
\usage{
generate_random_matrix(rows, cols)
}
\arguments{
\item{rows}{number of rows.}

\item{cols}{number of columns.}
}
\value{
a random matrix of dimension rows times columns with elements from
a standard normal distribution
}
\description{
Function to generate a random matrix with specified dimensions.
}
