% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{\%stk\%}
\alias{\%stk\%}
\title{The stk function.}
\usage{
nu \%stk\% mu
}
\arguments{
\item{nu}{neural network.}

\item{mu}{neural network.}
}
\value{
A stacked neural network of nu and mu.
}
\description{
The stk function.
}
