% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{hid}
\alias{hid}
\title{hid}
\usage{
hid(nu)
}
\arguments{
\item{nu}{a neural network of the type generated by create_nn()

By definition \eqn{\mathsf{H}(\nu) = \mathsf{D}(\nu) - 1}}
}
\value{
Integer representing the number of hidden layers.
}
\description{
The function that returns the number of hidden layers of a
neural network. Denoted \eqn{\mathsf{H}}
}
\examples{
create_nn(c(4, 5, 6, 2)) |> hid()

}
\references{
Definition 1.3.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}.
}
