% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{i}
\alias{i}
\title{i}
\usage{
i(d)
}
\arguments{
\item{d}{the size of the \eqn{\mathsf{i}} network}
}
\value{
returns the i_d network
}
\description{
The function that returns the \eqn{\mathbb{i}} network.
}
\examples{
i(5)
i(10)

}
\references{
Definition 2.2.6. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}
}
