% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Id.R
\name{Id}
\alias{Id}
\title{: Id}
\usage{
Id(d = 1)
}
\arguments{
\item{d}{the dimension of the \eqn{Id} network, by default it is \eqn{1}.}
}
\value{
Returns the \eqn{\mathsf{Id_1}} network.
}
\description{
The function that returns the \eqn{\mathsf{Id_1}} networks.
}
