% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiation.R
\name{inst}
\alias{inst}
\title{inst}
\usage{
inst(neural_network, activation_function, x)
}
\arguments{
\item{neural_network}{An ordered list of lists, of the type generated by
create_neural_network() where each element in the
list of lists is a pair \eqn{(W,b)} representing the weights and biases of
that layer.

\emph{NOTE:} We will call istantiation what Grohs et. al. call "realization".}

\item{activation_function}{A continuous function applied to the output of each layer}

\item{x}{our input to the continuous function formed from activation. Our input will
be an element in \eqn{\mathbb{R}^d} for some appropriate \eqn{d}.}
}
\value{
The output of the function that is the instantiation of the given
neural network with the given activation function at the given \eqn{x}.
}
\description{
The function that instantiates a neural network as created
by create_neural_network().
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al. Space-time error estimates for deep
neural network approximations for differential equations. Adv Comput Math 49, 4 (2023).
https://doi.org/10.1007/s10444-022-09970-2.

Definition 1.3.4. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}
}
