% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nn.R
\name{is_nn}
\alias{is_nn}
\title{is_nn}
\usage{
is_nn(nn)
}
\arguments{
\item{nn}{A neural network. Neural networks are defined to be an ordered
list of ordered pairs of \eqn{(W,b)}. Where \eqn{W} is the matrix
representing the weight matrix at that layer and \eqn{b} the bias vector.
Neural networks are defined to be elements belonging to the following set:
\deqn{
\mathsf{NN} = \bigcup_{L\in \N} \bigcup_{l_0,l_1,...,l_L \in \N}
\left( \times^L_{k=1} \left[ \mathbb{R}^{l_k \times l_{k-1}} \times \R^{l_k}\right]  \right)
}

See references for a more detailed discussion.}
}
\value{
TRUE or FALSE on whether nn is indeed a neural network as defined above.

We will use the definition of neural networks as found in the references section.#'
}
\description{
Function to create a list of lists for neural network layers
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al.
Space-time error estimates for deep neural network approximations
for differential equations. Adv Comput Math 49, 4 (2023).
\doi{https://rdcu.be/dxYVV}.
}
