% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{lay}
\alias{lay}
\title{lay}
\usage{
lay(nu)
}
\arguments{
\item{nu}{A neural network of the type generated by
create_nn(). Denoted \eqn{\mathsf{L}}.}
}
\value{
A tuple representing the layer architecture of our neural network.
}
\description{
The function that returns the layer architecture of a neural
network.
}
\examples{
create_nn(c(4, 5, 6, 2)) |> lay()
}
\references{
Definition 1.3.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}.
}
