% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_sum.R
\name{nn_sum}
\alias{nn_sum}
\alias{\%nn_sum\%}
\title{nn_sum}
\usage{
nn_sum(nu_1, nu_2)

nu_1 \%nn_sum\% nu_2
}
\arguments{
\item{nu_1}{A neural network.}

\item{nu_2}{A neural network.}
}
\value{
A neural network that is the neural network sum of \eqn{\nu_1} and \eqn{\nu_2}
i.e. \eqn{\nu_1 \oplus \nu_2}.

\emph{Note:} We have two versions, an infix version and a prefix version.
}
\description{
A function that performs the neural network sum for two
neural networks of the type generated by
create_neural_network(). Neural network sums are defined for
\eqn{\nu_1 \in \mathsf{NN}} and \eqn{\nu_2 \in \mathsf{NN}} as:
\deqn{
\oplus^v_{i=u}\nu_i \coloneqq \left( \mathsf{Sum}_{v-u+1,\mathsf{O}(\nu_2)} \bullet \left[ \boxminus^v_{i=u}\nu_i \right] \bullet \mathsf{Cpy}_{(v-u+1),\mathsf{I}(\nu_1)} \right)
}
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al.
Space-time error estimates for deep neural network approximations
for differential equations. Adv Comput Math 49, 4 (2023).
\doi{https://rdcu.be/dxYVV}.
}
