% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_nner.R
\name{view_nn}
\alias{view_nn}
\title{view_nner}
\usage{
view_nn(nn)
}
\arguments{
\item{nn}{A neural network of the type creted by \code{\link{create_nn()}}, i.e.
a list of lists of \eqn{W} and \eqn{b}.}
}
\value{
A displayed version of the neural network. This may be required if
the neural network is very deep.
}
\description{
view_nner
}
\examples{
Xpn(3,2.1,1.1) |> view_nn()
Pwr(2,2.1,0.1) |> view_nn()
}
