% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nrm.R
\name{Nrm}
\alias{Nrm}
\title{Nrm}
\usage{
Nrm(d)
}
\arguments{
\item{d}{the dimensions of the vector being normed.}
}
\value{
a neural network that takes the 1-norm of a vector of
size d.under ReLU activation.

\emph{Note:} This function is split into two cases
much like the definition itself.

For a specific definition, see:
}
\description{
A function that creates the \eqn{\mathsf{Nrm}} neural networks.that take
the 1- norm of a \eqn{d}-dimensional vector when instantiated with ReLU
activation.
}
\references{
Lemma 4.2.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}
}
