% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{out}
\alias{out}
\title{out}
\usage{
out(nu)
}
\arguments{
\item{nu}{A neural network of the type generated by create_nn().}
}
\value{
An integer representing the output width of the neural network.
}
\description{
The function that returns the output layer size of a neural
network. Denoted \eqn{\mathsf{O}}.
}
\examples{
create_nn(c(4, 5, 6, 2)) |> out()
}
\references{
Definition 1.3.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}.
}
