% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.R
\name{P}
\alias{P}
\title{The P neural network}
\usage{
P(X, y, L)
}
\arguments{
\item{X}{a matrix. More precisely a collection of n row vectors of size d.
That is to say an d by n matrix}

\item{y}{a row vector of size m that is the ouput of the function at each
of the n sample points}

\item{L}{the Lipschitz constant for the function.}
}
\value{
A neural network that gives the maximum convolution approximation
of a function whose outputs is \eqn{y} at \eqn{n} sample points given by
each row of \eqn{X}, when instantiated with ReLU.
}
\description{
The P neural network
}
\references{
Lemma 4.2.9. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}.
}
