% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi.R
\name{Phi}
\alias{Phi}
\title{The Phi function}
\usage{
Phi(eps)
}
\arguments{
\item{eps}{parameter for Phi}
}
\value{
neural network Phi that approximately squares a number between
0 and 1.
}
\description{
The Phi function
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al.
Space-time error estimates for deep neural network approximations
for differential equations. Adv Comput Math 49, 4 (2023).
https://doi.org/10.1007/s10444-022-09970-2
}
