% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_k.R
\name{Phi_k}
\alias{Phi_k}
\title{The Phi_k function}
\usage{
Phi_k(k)
}
\arguments{
\item{k}{an integer \eqn{k \in (2,\infty)}}
}
\value{
The Phi_k neural network
}
\description{
The Phi_k function
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al.
Space-time error estimates for deep neural network approximations
for differential equations. Adv Comput Math 49, 4 (2023).
https://doi.org/10.1007/s10444-022-09970-2
}
