% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prd.R
\name{Prd}
\alias{Prd}
\title{Prd}
\usage{
Prd(q, eps)
}
\arguments{
\item{q}{A real number within \eqn{(2,\infty)}}

\item{eps}{A real number within \eqn{(0,\infty)}}
}
\value{
A neural network that takes in \eqn{x} and \eqn{y} and approximately
returns \eqn{xy} when instantiated with ReLU activation, and given a list
c(x,y), the two numbers to be multiplied.
}
\description{
A function that returns the \eqn{\mathsf{Prd}} neural networks that
approximates the product of two real numbers when given an appropriate
\eqn{q}, \eqn{\varepsilon}, a real number \eqn{x} and instantiation with ReLU.
activation.
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al. Space-time error estimates for deep
neural network approximations for differential equations. Adv Comput Math 49, 4 (2023).
https://doi.org/10.1007/s10444-022-09970-2
}
