% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pwr.R
\name{Pwr}
\alias{Pwr}
\title{Pwr}
\usage{
Pwr(q, eps, exponent)
}
\arguments{
\item{q}{inside \eqn{(2,\infty)}.}

\item{eps}{inside \eqn{(0,\infty)}.}

\item{exponent}{the exponent which the Pwr network will approximate. Must be
a non-negative integer.}
}
\value{
A neural network that approximates raising a number to exponent, when
given appropriate \eqn{q,\varepsilon} and exponent when isntanatiated
under ReLU activation at \eqn{x}.
}
\description{
A function that returns the \eqn{\mathsf{Pwr}} neural networks.
}
