% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{ReLU}
\alias{ReLU}
\title{: ReLU}
\usage{
ReLU(x)
}
\arguments{
\item{x}{A real number that is the input to our ReLU function.}
}
\value{
The output of the standard ReLU function, i.e. \eqn{\max\{0,x\}}. See also \code{\link{Sigmoid}}.
and \code{\link{Tanh}}.
}
\description{
The ReLU activation function
}
\examples{
ReLU(5)
ReLU(-5)

}
