% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{Sigmoid}
\alias{Sigmoid}
\title{: Sigmoid}
\usage{
Sigmoid(x)
}
\arguments{
\item{x}{a real number that is the input to our Sigmoid function.}
}
\value{
The output of a standard Sigmoid function,
i,e. \eqn{\frac{1}{1 + \exp(-x)}}.
See also \code{\link{Tanh}}.and \code{\link{ReLU}}.
}
\description{
The Sigmoid activation function.
}
\examples{
Sigmoid(0)
Sigmoid(-1)

}
