% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sne.R
\name{Sne}
\alias{Sne}
\title{Sne}
\usage{
Sne(n, q, eps)
}
\arguments{
\item{n}{the number of Taylor iterations}

\item{q}{a real number in \eqn{(2,\infty)}}

\item{eps}{a real number in \eqn{(0,\infty)}

\emph{Note: } In practice for most desktop uses
\eqn{q < 2.05} and \eqn{\varepsilon< 0.05} tends to cause problems in
"too long a vector", atleaast as tested on my computer.}
}
\value{
A neural network that approximates \eqn{\sin} when given
an appropriate \eqn{n,q,\varepsilon} and instantiated with ReLU
activation and given value \eqn{x}.
}
\description{
Returns the \eqn{\mathsf{Sne}} neural networks
}
