% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sqr.R
\name{Sqr}
\alias{Sqr}
\title{Sqr}
\usage{
Sqr(q, eps)
}
\arguments{
\item{q}{parameter for the Sqr network. \eqn{2 \in (2,\infty)}}

\item{eps}{parameter for the Sqr network. \eqn{eps \in (0,1]}. You may
choose epsilon to be greater than 1 but that leads to large errors}
}
\value{
A neural network that approximates the square of a real number.when
provided appropriate \eqn{q,\varepsilon} and upon instantiation with ReLU,
and a real number \eqn{x}
}
\description{
A function that returns the \eqn{\mathsf{Sqr}} neural networks.
}
\references{
Grohs, P., Hornung, F., Jentzen, A. et al. Space-time error estimates for deep
neural network approximations for differential equations. Adv Comput Math 49, 4 (2023).
https://doi.org/10.1007/s10444-022-09970-2
}
