% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_mult.R
\name{srm}
\alias{srm}
\alias{\%srm\%}
\title{srm}
\usage{
srm(nu, a)

nu \%srm\% a
}
\arguments{
\item{nu}{A neural network}

\item{a}{A real number.}
}
\value{
Returns a neural network that is \eqn{\nu \triangleleft a}. This
instantiates as \eqn{f(a \cdot x)}.under continuous function activation. More
specifically we will define this operation as:

Let \eqn{\lambda \in \mathbb{R}}. We will denote by \eqn{(\cdot) \triangleleft (\cdot):
\mathsf{NN} \times \mathbb{R} \rightarrow \mathsf{NN}} the function satisfying for all
\eqn{\nu \in \mathsf{NN}} and \eqn{\lambda \in \mathbb{R}} that \eqn{\nu \triangleleft \lambda =
\nu \bullet \mathsf{Aff}_{\lambda \mathbb{I}_{\mathsf{I}(\nu)},0}}.
}
\description{
The function that returns the right scalar multiplication
neural network
}
\references{
Definition 2.3.4. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}.

\emph{Note:} We will have two versions of this operation, a prefix
and an infix version.
}
