% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aff.R
\name{Sum}
\alias{Sum}
\title{Sum}
\usage{
Sum(n, k)
}
\arguments{
\item{n}{number of copies of a certain vector to be summed.}

\item{k}{the size of the summation vector.}
}
\value{
An affine neural network that will take a vector of size
\eqn{n \times k} and return the summation vector that is of length
\eqn{k}. See also \code{\link{Aff}} and \code{\link{Cpy}}.
}
\description{
The function that returns \eqn{\mathsf{Sum}} neural networks.

These are neural networks defined as such
\deqn{
\mathsf{Aff}_{\left[ \mathbb{I}_k \: \mathbb{I}_k \: \cdots \: \mathbb{I}_k\right],0_{k}}
}
}
\references{
Definition 2.4.1. Jentzen, A., Kuckuck, B., and von Wurstemberger, P. (2023).
Mathematical introduction to deep learning: Methods, implementations,
and theory. \url{https://arxiv.org/abs/2310.20360}
}
