% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{Tanh}
\alias{Tanh}
\title{Tanh}
\usage{
Tanh(x)
}
\arguments{
\item{x}{a real number}
}
\value{
the \eqn{tanh} of x. See also \code{\link{Sigmoid}} and
\code{\link{ReLU}}.
}
\description{
The tanh activation function
}
\examples{
Tanh(0)
Tanh(0.1)

}
