% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etr.R
\name{Trp}
\alias{Trp}
\title{Trp}
\usage{
Trp(h)
}
\arguments{
\item{h}{the horizontal distance between two mesh points}
}
\value{
The \eqn{\mathsf{Trp}} network that gives the area
when activated with ReLU and two meshpoint values x_1 and x_2.
}
\description{
The function that returns the \eqn{\mathsf{Trp}} networks.
}
