% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_viewer.R
\name{view_nn}
\alias{view_nn}
\title{view_nn}
\usage{
view_nn(nn)
}
\arguments{
\item{nn}{A neural network., i.e.
a list of lists of \eqn{W} and \eqn{b}.}
}
\value{
A displayed version of the neural network. This may be required if
the neural network is very deep.
}
\description{
Takes a neural network shown in vectorized form and explicitly
displays it.
}
\examples{
c(5, 6, 7, 9) |>
  create_nn() |>
  view_nn()
Sqr(2.1, 0.1) |> view_nn()

Xpn(3, 2.1, 1.1) |> view_nn()
Pwr(2.1, 0.1, 3) |> view_nn()
}
