% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_k.R
\docType{data}
\name{A}
\alias{A}
\title{This is an intermediate variable. See the reference}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 1 columns.
}
\usage{
A
}
\description{
This is an intermediate variable. See the reference
}
\references{
Definition 2.22. Rafi S., Padgett, J.L., Nakarmi, U. (2024)
Towards an Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
\keyword{datasets}
