% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Id.R
\name{Id}
\alias{Id}
\title{: Id}
\usage{
Id(d = 1)
}
\arguments{
\item{d}{the dimension of the \eqn{Id} network, by default it is \eqn{1}.}
}
\value{
Returns the \eqn{\mathsf{Id_1}} network.
}
\description{
The function that returns the \eqn{\mathsf{Id_1}} networks.
}
\examples{
Id()
Id(3)

}
\references{
Definition 2.17. Rafi S., Padgett, J.L., Nakarmi, U. (2024)
Towards an Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
