% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi.R
\name{Phi}
\alias{Phi}
\title{The Phi function}
\usage{
Phi(eps)
}
\arguments{
\item{eps}{parameter for Phi in \eqn{(0,\infty)}}
}
\value{
neural network Phi that approximately squares a number between
0 and 1.
}
\description{
The Phi function
}
\examples{
Phi(0.5) |> view_nn()
Phi(0.1) |> view_nn()

}
\references{
Definition 2.23. Rafi S., Padgett, J.L., Nakarmi, U. (2024)
Towards an Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
