% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_k.R
\name{Phi_k}
\alias{Phi_k}
\title{The Phi_k function}
\usage{
Phi_k(k)
}
\arguments{
\item{k}{an integer \eqn{k \in (2,\infty)}}
}
\value{
The Phi_k neural network
}
\description{
The Phi_k function
}
\examples{
Phi_k(4) |> view_nn()
Phi_k(5) |> view_nn()

}
\references{
Definition 2.22. Rafi S., Padgett, J.L., Nakarmi, U. (2024)
Towards an Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
