% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pwr.R
\name{Pwr}
\alias{Pwr}
\title{Pwr}
\usage{
Pwr(q, eps, exponent)
}
\arguments{
\item{q}{a real number in \eqn{(2,\infty)}. Accuracy as well as computation
time increases as \eqn{q} gets closer to \eqn{2} increases}

\item{eps}{a real number in \eqn{(0,\infty)}. ccuracy as well as computation
time increases as \eqn{\varepsilon} gets closer to \eqn{0} increases}

\item{exponent}{The power to which we will raise. Computation
time increases as exponent increases}
}
\value{
A neural network that approximates raising a number to exponent, when
given appropriate \eqn{q,\varepsilon} and exponent when instantiated
under ReLU activation at \eqn{x}.
}
\description{
A function that returns the \eqn{\mathsf{Pwr}} neural networks.
}
\examples{
Pwr(2.1, 0.1, 2) |> inst(ReLU, 3) # This may take some time, please only click once.

}
