% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etr.R
\name{Trp}
\alias{Trp}
\title{Trp}
\usage{
Trp(h)
}
\arguments{
\item{h}{the horizontal distance between two mesh points}
}
\value{
The \eqn{\mathsf{Trp}} network that gives the area
when activated with ReLU or any continuous function and two
meshpoint values \eqn{x_1} and \eqn{x_2}.
}
\description{
The function that returns the \eqn{\mathsf{Trp}} networks.
}
\examples{
Trp(0.1)
Trp(0.5) |> inst(ReLU, c(9, 7))
Trp(0.1) |> inst(Sigmoid, c(9, 8))

}
\references{
Definition 2.31. Rafi S., Padgett, J.L., Nakarmi, U. (2024)
Towards an Algebraic Framework For
Approximating Functions Using Neural Network Polynomials
\url{https://arxiv.org/abs/2402.01058}
}
